added a new Event component that displayed the information about an event
